Add rail-drawing macros to square_corner_twist_borders
These macros encapsulate a relatively easy construction that is
annoying to have to figure out every time one wants to add the
rails to the border design. The provided macros generate paths
for the upper right horizontal and vertical rails. One of the
sample figures shows how to extend to all four quadrants.